Module Drom_toml.Printer

val value : Stdlib.Format.formatter -> Types.value -> unit

Given a Toml value and a formatter, inserts a valid Toml representation of this value in the formatter.

val table : Stdlib.Format.formatter -> Types.table -> unit

Given a Toml table and a formatter, inserts a valid Toml representation of this value in the formatter.

val array : Stdlib.Format.formatter -> Types.array -> unit

Given a Toml array and a formatter, inserts a valid Toml representation of this value in the formatter.

raises Invalid_argument
if the array is an array of tables

val string_of_value : Types.value -> string

Turns a Toml value into a string.

val string_of_table : Types.table -> string

Turns a Toml table into a string.

val string_of_array : Types.array -> string

Turns a Toml array into a string.